Created BaseActivity that implements ad free upgrade, applied to DisplayActivity
