+++ app/src/main/java/com/a494studios/koreanconjugator/utils/BaseActivity.java
@@ -1,0 +1,66 @@
+package com.a494studios.koreanconjugator.utils;
+
+import android.app.SearchManager;
+import android.content.Intent;
+import android.view.Menu;
+import android.view.MenuInflater;
+import android.view.MenuItem;
+
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.SearchView;
+
+import com.a494studios.koreanconjugator.R;
+import com.a494studios.koreanconjugator.display.DisplayActivity;
+import com.a494studios.koreanconjugator.settings.SettingsActivity;
+
+import org.rm3l.maoni.Maoni;
+
+import static com.eggheadgames.aboutbox.activity.AboutActivity.launch;
+
+public abstract class BaseActivity extends AppCompatActivity {
+
+    private SearchView searchView;
+
+    @Override
+    public boolean onCreateOptionsMenu(Menu menu) {
+        MenuInflater inflater = getMenuInflater();
+        inflater.inflate(R.menu.search_menu, menu);
+        SearchManager searchManager = (SearchManager) getSystemService(SEARCH_SERVICE);
+        searchView = (SearchView) menu.findItem(R.id.menu_search).getActionView();
+        searchView.setSearchableInfo(searchManager.getSearchableInfo(getComponentName()));
+        return true;
+    }
+
+    @Override
+    public boolean onOptionsItemSelected(MenuItem item) {
+        if(item.getItemId() == R.id.overflow_settings){
+            startActivity(new Intent(getBaseContext(), SettingsActivity.class));
+            return true;
+        }else if(item.getItemId() == R.id.overflow_about){
+            Utils.makeAboutBox(this);
+            launch(this);
+            return true;
+        }else if(item.getItemId() == R.id.overflow_bug) {
+            Maoni maoni = Utils.makeMaoniActivity(this);
+            if (maoni != null) {
+                maoni.start(this);
+            }
+            return true;
+        } else if(item.getItemId() == R.id.overflow_ad_free) {
+            Utils.showAdFreeUpgrade(this);
+            return true;
+        } else {
+            return super.onOptionsItemSelected(item);
+        }
+    }
+
+    @Override
+    public void onResume() {
+        super.onResume();
+        if(searchView != null) {
+            searchView.setQuery("", false);
+            searchView.setIconified(true);
+            searchView.clearFocus();
+        }
+    }
+}
